Fine-tuned some audio features.
Found out that windows PlaySound can only play a single sound at once in a program. Need to find an audio solution for multi-channel output. I've been able to link OpenAL but I can't figure out how to actually load wav files.

Fmod is a nightmare to deal with and SFML won't even compile properly with Codeblocks 17.12. Not sure what to do. Might just stick with the single channel and just play the guitar chords on it.