IrrKlang installed and working
AudioSource has been updated to work with IrrKlang audio library.

Can now utilize multiple audio sources as well as positional audio for applying distance to the volume of sound.

Still have a lot of work to do on this front but I can start implementing my algorithm for chord generation.