Implemented Basic User Input For Chords
If user holds left shift, a left or right mouse icon will show up next to the player on the left or right side, accordingly. If user presses the correct input while it's displayed, the input is considered successful.

The input randomizes every two seconds if user has not pressed a mouse button while shift is held.

Need to implement chord playing now.